+++ CHANGES.md
@@ -20,7 +20,7 @@
 ## [WIP] 0.7.1 / 2017-July-??
 
 - [TODO] Fix issue: Validator is not applied when (k config not-found) arity is used
-- [TODO] Fix issue: Validator is not applied in `lookup-keypath`
+- Fix issue: Value parser is not applied in `lookup-keypath`
 
 
 ## 0.7.0 / 2017-July-30